[clean] rename to psfnetlens
